refactor(web): extract usePresets hook from App.jsx (sc-1651)
Second Phase-A slice. Moves recipe-preset state + presetQuery, refreshPresets,
createPreset, updatePreset, duplicatePreset, deletePreset into
hooks/usePresets.js. App composes it ({ token, activeProject, setError }) and
forwards the returned values down unchanged, so PresetManagerScreen/ImageStudio/
VideoStudio props are untouched. App's bulk refreshData still seeds the list via
the hook's setPresets (same React setter identity). presetQuery stays internal to
the hook (only the preset mutations used it).

Behavior-preserving: App.jsx 2186 -> 2132 lines. 107 web vitest pass, build clean,
browser smoke (Rust API on a temp data dir) confirmed the Preset Manager renders,
refreshPresets loads a seeded preset, and Duplicate round-trips through the hook
(2 presets) with no console errors.
